fix issue with multiple instances of the same subtitle in result file.
The root cause of the bug is that vob_t contains a packet for every packet
in the original mpeg stream. If a subtitle image is encoded in multiple
mpeg packets, vob_t will contain multiple packets pointing to the same
subtitle.
Fixed the issue by only processing the first packet and skip all subsequent
packets of the same subtitle.